[Starts #162350735] Rename instances of models package to incidence and update its usage
